+++ test/test_runner.c
@@ -617,7 +617,7 @@ static u32 FunctionTest_RandomUniform(enum RandomTag tag, u32 lo, u32 hi, bool32
     return RandomUniformDefaultValue(tag, lo, hi, reject, caller);
 }
 
-static u32 FunctionTest_RandomWeightedArray(enum RandomTag tag, u32 sum, u32 n, const u8 *weights, void *caller)
+static u32 FunctionTest_RandomWeightedArray(enum RandomTag tag, u32 sum, u32 n, const u16 *weights, void *caller)
 {
     //rigged
     for (u32 i = 0; i < RIGGED_RNG_COUNT; i++)
@@ -1030,7 +1030,7 @@ u32 RandomUniformExcept(enum RandomTag tag, u32 lo, u32 hi, bool32 (*reject)(u32
         return RandomUniformExceptDefault(tag, lo, hi, reject);
 }
 
-u32 RandomWeightedArray(enum RandomTag tag, u32 sum, u32 n, const u8 *weights)
+u32 RandomWeightedArray(enum RandomTag tag, u32 sum, u32 n, const u16 *weights)
 {
     void *caller = __builtin_extract_return_addr(__builtin_return_address(0));
     if (gTestRunnerState.test->runner->randomWeightedArray)
@@ -1063,7 +1063,7 @@ u32 RandomUniformDefaultValue(enum RandomTag tag, u32 lo, u32 hi, bool32 (*rejec
     return default_;
 }
 
-u32 RandomWeightedArrayDefaultValue(enum RandomTag tag, u32 n, const u8 *weights, void *caller)
+u32 RandomWeightedArrayDefaultValue(enum RandomTag tag, u32 n, const u16 *weights, void *caller)
 {
     while (weights[n-1] == 0)
     {
